+++ src/app/shared/settings.ts
@@ -1,6 +1,7 @@
 export enum Settings {
     Path = 'todoFilePath',
     TaskFilter = 'taskFiler',
+    TaskOrdering = 'taskOrdering',
 }
 
 export interface TaskFilter {